# src/security-server/openapi-model/src/main/resources/META-INF/openapi-definition.yaml
...
openapi: 3.0.0
info:
  description: X-Road Security Server Admin API. Note that the error metadata responses described in some endpoints are subjects to change and may be updated in upcoming versions.
  version: "2.1.0"
  title: X-Road Security Server Admin API
  contact:
    name: Nordic Institute for Interoperability Solutions (NIIS)
    url: https://github.com/nordic-institute/X-Road-development/#enhancement-requests-and-error-reports
    email: info@niis.org
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: basepath for API v1
...
paths:
...
  /member-classes:
    get:
      tags:
        - member-classes
      summary: get list of known member classes
      operationId: getMemberClasses
      description: <h3>Administrator lists member classes.</h3>
      parameters:
        - in: query
          name: current_instance
          description: if true, return member classes for this instance.
            if false (default), return member classes for all instances
          required: false
          schema:
            default: false
            type: boolean
      responses:
        '200':
          description: array of member classes
          content:
            application/json:
              schema:
                type: array
                description: array of member classes
                uniqueItems: true
                items:
                  type: string
                  format: text
                  minLength: 1
                  maxLength: 255
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
  /member-classes/{id}:
    get:
      tags:
        - member-classes
      summary: get list of known member classes for a given instance
      operationId: getMemberClassesForInstance
      description: <h3>Administrator lists member classes for a given instance.</h3>
      parameters:
        - in: path
          name: id
          description: instance id
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 255
      responses:
        '200':
          description: array of member classes
          content:
            application/json:
              schema:
                type: array
                description: array of member classes
                uniqueItems: true
                items:
                  type: string
                  format: text
                  minLength: 1
                  maxLength: 255
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: X-Road-ApiKey token=<api key>
      # https://github.com/nordic-institute/X-Road-REST-UI/blob/XRDDEV-237/doc/Manuals/ug-ss_x-road_6_security_server_user_guide.md#1911-creating-new-api-keys
...
security:
  - ApiKeyAuth: []